pyproject.toml
[project]
name = "japanese-marketplace-scraper"
version = "0.1.0"
description = "A Python scraper for Japanese marketplaces to find lowest price products"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "pandas>=2.0.0",
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    "fake-useragent>=1.4.0",
    "lxml>=4.9.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "packaging>=25.0",
    "setuptools>=80.9.0",
    # Dashboard dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "openai>=1.97.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
